added alternate way of retrieving aws db info using aws_secrets_manager.py
